Got some button handling working

+++ query.go
@@ -3,5 +3,5 @@ package gowave
 type Query struct {
 	ID          uint8
 	Type        uint8
-	PayloadSize uint16
+	PayloadSize int
 }
